add movie and products section -restructure folders -restructure codes
