+++ src/api/index.js
@@ -1,0 +1,5 @@
+import axios from 'axios';
+
+export const _2gedaservice = axios.create({
+    baseURL: "https://development.2geda.net"
+})
