+++ src/components/Dashboard/Stick.jsx
@@ -1,4 +1,5 @@
 import { useEffect, useState } from "react";
+import PostImage from "assets/images/sample-avatar.png"
 
 const Stick = (address, username, cover_image) => {
   const [isClicked, setIsClicked] = useState(false);
@@ -18,15 +19,10 @@ const Stick = (address, username, cover_image) => {
   return (
     <div className="stick-cont">
       <div className="post-profile">
-        {/* {cover_image && cover_image.cover_image ? (
-          <img src={cover_image.cover_image} />
-        ) : (
-          <img src="https://dummyimage.com/300.png/09f/fff&text=Ash+Allen" />
-        )} */}
-        <img src="https://dummyimage.com/300.png/09f/fff&text=Ash+Allen" alt="dum"/>
+        <img src={PostImage} alt="dum"/>
 
         <div className="post-profile-details">
-          <div className="post-profile-name"> {username}</div>
+          <div className="post-profile-name"> Random name</div>
           <div className="autor-ooby">Pharmacist</div>
           <div className="autor-location">Lagos, Nigeria</div>
         </div>
