+++ src/components/Dashboard/second.css
@@ -70,23 +70,32 @@
   align-items: center;
   justify-content: center;
 }
+
+.tic-btn-con a {
+  text-decoration: none;
+  font-size: 1rem;
+}
+
 .tick-btn {
   width: 80px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   border: none;
   outline: none;
   background: #4f0da3;
   color: #fff;
-  height: 30px;
+  height: 26px;
   border-radius: 8px;
   cursor: pointer;
 }
 .pro-title {
-  font-size: 20px;
+  font-size: 1.1rem;
   font-weight: 700;
   padding-left: 5px;
 }
 .rate-con {
-  font-size: 16px;
+  font-size: 1rem;
   display: flex;
   align-items: center;
   gap: 5px;
@@ -110,10 +119,10 @@
   font-size: 13px;
 }
 .pro-btn-con {
-  font-size: 20px;
+  font-size: 1.2rem;
   font-weight: 500;
   padding-left: 5px;
-  margin-bottom: 10px;
+  margin-bottom: .3rem;
 }
 
 .movie-das-card-box {
@@ -178,92 +187,6 @@
   justify-content: space-between;
   margin-top: 15px;
 }
-.swiper-container {
-  margin-top: 30px;
-}
-.swiper_container {
-  height: 52rem;
-  position: relative;
-  margin-bottom: -14em;
-}
-
-.swiper-slide {
-  height: 42rem;
-  position: relative;
-}
-
-.swiper-slide img {
-  width: 20em;
-  height: 280px;
-  border-radius: 2rem;
-  object-fit: cover;
-}
-
-.swiper-slide-shadow-left,
-.swiper-slide-shadow-right {
-  display: none;
-}
-
-.slider-controler {
-  position: relative;
-  bottom: 2rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  display: none;
-}
-
-.slider-controler .swiper-button-next {
-  left: 58% !important;
-  transform: translateX(-58%) !important;
-}
-
-/* @media (max-width: 990px) {
-    .slider-controler .swiper-button-next {
-      left: 70% !important;
-      transform: translateX(-70%) !important;
-    }
-  }
-  
-  @media (max-width: 450px) {
-    .slider-controler .swiper-button-next {
-      left: 80% !important;
-      transform: translateX(-80%) !important;
-    }
-  }
-  
-  @media (max-width: 990px) {
-    .slider-controler .swiper-button-prev {
-      left: 30% !important;
-      transform: translateX(-30%) !important;
-    }
-  }
-  
-  @media (max-width: 450px) {
-    .slider-controler .swiper-button-prev {
-      left: 20% !important;
-      transform: translateX(-20%) !important;
-    }
-  } */
-
-.slider-controler .slider-arrow {
-  background: var(--white);
-  width: 3.5rem;
-  height: 3.5rem;
-  border-radius: 50%;
-  left: 42%;
-  transform: translateX(-42%);
-  filter: drop-shadow(0px 8px 24px rgba(18, 28, 53, 0.1));
-}
-
-.slider-controler .slider-arrow ion-icon {
-  font-size: 2rem;
-  color: #222224;
-}
-
-.slider-controler .slider-arrow::after {
-  content: "";
-}
 
 .swiper-pagination {
   position: relative;
@@ -277,41 +200,6 @@
 
 .swiper-pagination .swiper-pagination-bullet-active {
   background: var(--primary);
-}
-.im-cont-box {
-  position: relative;
-  width: 100%;
-}
-.im-cont-box .mid-lo {
-  width: 40px;
-  height: 40px;
-}
-.mid-lo {
-  position: absolute;
-  z-index: 111;
-  top: 40%;
-  left: 50%;
-}
-.slid-conb {
-  position: absolute;
-  bottom: 0.5%;
-  z-index: 1111;
-  color: #fff;
-  background: rgba(0, 0, 0, 0.3);
-  width: 20em;
-  padding: 10px;
-  border-bottom-left-radius: 2rem;
-  border-bottom-right-radius: 2rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.slid-text-main {
-  font-size: 12px;
-}
-.kemb {
-  font-size: 20px;
-  cursor: pointer;
 }
 
 /* Search Section */
@@ -389,39 +277,6 @@
   display: none;
 }
 
-@media (max-width: 500px) {
-  .swiper_container {
-    height: 47rem;
-  }
-  .swiper-slide {
-    width: 28rem !important;
-    height: 36rem !important;
-  }
-  .swiper-slide img {
-    width: 30rem !important;
-  }
-  .im-cont-box .mid-lo {
-    width: 40px !important;
-    height: 40px !important;
-  }
-  .mid-lo {
-    top: 40%;
-    left: 45% !important;
-  }
-  .slid-conb {
-    width: 30em !important;
-  }
-  .slid-text-main {
-    font-size: 15px;
-  }
-  .kemb {
-    font-size: 25px;
-    cursor: pointer;
-  }
-  .slid-text {
-    font-size: 12px;
-  }
-}
 @media only screen and (max-width: 1150px) {
   .general-search-container {
     right: 10%;
