+++ src/components/Dashboard/style.css
@@ -160,7 +160,7 @@
 .over-lay-post {
   position: absolute;
   top: 0;
-  background: #403f3f77;
+  background: #4949493a;
   width: 100%;
   height: 100%;
   color: #fff;
