+++ src/components/Home/Movieslider/movieslide.css
@@ -1,0 +1,151 @@
+.swiper-container {
+	margin-top: 2rem;
+}
+
+.swiper-slide img {
+	width: 17em;
+	height: 23em;
+	border-radius: 2rem;
+	object-fit: cover;
+}
+
+.slid-conb {
+	position: absolute;
+	bottom: 0.5%;
+	z-index: 1111;
+	color: #fff;
+	background: rgba(0, 0, 0, 0.3);
+    width: inherit;
+	width: 17em;
+	padding: 10px;
+	border-bottom-left-radius: 2rem;
+	border-bottom-right-radius: 2rem;
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+}
+
+.swiper-slide-shadow-left,
+.swiper-slide-shadow-right {
+	display: none;
+}
+
+.slider-controler {
+	position: relative;
+	bottom: 2rem;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	display: none;
+}
+
+.slider-controler .swiper-button-next {
+	left: 58% !important;
+	transform: translateX(-58%) !important;
+}
+
+.slider-controler .slider-arrow {
+	background: var(--white);
+	width: 3.5rem;
+	height: 3.5rem;
+	border-radius: 50%;
+	left: 42%;
+	transform: translateX(-42%);
+	filter: drop-shadow(0px 8px 24px rgba(18, 28, 53, 0.1));
+}
+
+.slider-controler .slider-arrow ion-icon {
+	font-size: 2rem;
+	color: #222224;
+}
+
+.slider-controler .slider-arrow::after {
+	content: "";
+}
+
+.slid-text-main {
+	font-size: 12px;
+}
+.kemb {
+	font-size: 20px;
+	cursor: pointer;
+}
+
+.im-cont-box {
+	position: relative;
+	width: 100%;
+}
+
+.im-cont-box .mid-lo {
+	width: 40px;
+	height: 40px;
+}
+
+.mid-lo {
+	position: absolute;
+	z-index: 111;
+	top: 40%;
+	left: 50%;
+}
+
+@media (max-width: 500px) {
+	.swiper_container {
+		height: 32rem;
+	}
+	.swiper-slide {
+		width: 28rem !important;
+		height: 36rem !important;
+	}
+	.swiper-slide img {
+		width: 30rem !important;
+	}
+	.im-cont-box .mid-lo {
+		width: 40px !important;
+		height: 40px !important;
+	}
+	.mid-lo {
+		top: 40%;
+		left: 45% !important;
+	}
+	.slid-conb {
+		width: 30em !important;
+	}
+	.slid-text-main {
+		font-size: 15px;
+	}
+	.kemb {
+		font-size: 25px;
+		cursor: pointer;
+	}
+	.slid-text {
+		font-size: 12px;
+	}
+}
+
+/* @media (max-width: 990px) {
+    .slider-controler .swiper-button-next {
+      left: 70% !important;
+      transform: translateX(-70%) !important;
+    }
+  }
+  
+  @media (max-width: 450px) {
+    .slider-controler .swiper-button-next {
+      left: 80% !important;
+      transform: translateX(-80%) !important;
+    }
+  }
+  
+  @media (max-width: 990px) {
+    .slider-controler .swiper-button-prev {
+      left: 30% !important;
+      transform: translateX(-30%) !important;
+    }
+  }
+  
+  @media (max-width: 450px) {
+    .slider-controler .swiper-button-prev {
+      left: 20% !important;
+      transform: translateX(-20%) !important;
+    }
+  } */
